Handouts-Textbook-Research Methods An Engineering Approach
A PhD or master’s level research project is an enormous undertaking, and you might find yourself a bit uncertain about the process or how to achieve the desired outcome. In this research course, you will learn the underlying principles that are needed to conduct research from an engineering perspective. This course is designed for engineering students conducting postgraduate research work on engineering projects. The objective of the course is to translate current research methods, which are mostly from a social science perspective, into something more relatable and understandable to engineers. Our hope for this course is to go beyond the concepts to understand the actual reasons for doing research in a certain way. While engineers are the main target audience, non-engineers will find this information useful as well. The methods taught in this course will equip you with the knowledge needed to design, plan and construct your own research process